requirements.txt
ai-edge-litert
httpx
jaxtyping
gradio
gradio_client
fastapi
uvicorn
langchain
langchain-google-genai
langgraph
numpy
matplotlib
openpyxl
pandas
pydantic
python-dotenv
PyYAML
scikit-learn
scipy
SQLAlchemy
tqdm
tensorflow==2.19.0
